src/requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
channels>=4.1
channels-redis>=4.1
daphne>=4.1
paho-mqtt>=2.0
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
python-decouple>=3.8
django-cors-headers>=4.3
